backend/requirements.txt
Flask==3.1.0
Flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
flask-marshmallow==1.3.0
marshmallow-sqlalchemy==1.4.1
webargs==8.6.0
werkzeug==3.1.3
python-dotenv==1.0.1
requests==2.32.3
bcrypt==4.3.0